# ast, topdown: Improve indexer to understand function args