--- a/ast/compile.go
+++ b/ast/compile.go
@@ -1162,7 +1162,7 @@ func (c *Compiler) rewriteExprTerms() {
 
 // rewriteTermsInHead will rewrite rules so that the head does not contain any
 // terms that require evaluation (e.g., refs or comprehensions). If the key or
-// value contains or more of these terms, the key or value will be moved into
+// value contains one or more of these terms, the key or value will be moved into
 // the body and assigned to a new variable. The new variable will replace the
 // key or value in the head.
 //