--- a/ast/compile_test.go
+++ b/ast/compile_test.go
@@ -2284,7 +2284,7 @@ func TestRewriteDeclaredVars(t *testing.T) {
 					f(input, "bar")
 				}
 
-				f(x,y) {
+				f(x, y) {
 					x[y]
 				}
 				`,